+++ zetaclient/chains/evm/observer/v2_inbound.go
@@ -340,9 +340,6 @@ func (ob *Observer) parseAndValidateCallEvents(
 
 // newCallInboundVote creates a MsgVoteInbound message for a Call event
 func (ob *Observer) newCallInboundVote(event *gatewayevm.GatewayEVMCalled) types.MsgVoteInbound {
-	// determine confirmation mode
-	confirmationMode := ob.GetInboundConfirmationMode(event.Raw.BlockNumber)
-
 	return *types.NewMsgVoteInbound(
 		ob.ZetacoreClient().GetKeys().GetOperatorAddress().String(),
 		event.Sender.Hex(),
@@ -361,7 +358,7 @@ func (ob *Observer) newCallInboundVote(event *gatewayevm.GatewayEVMCalled) types
 		types.ProtocolContractVersion_V2,
 		false, // currently not relevant since calls are not arbitrary
 		types.InboundStatus_SUCCESS,
-		confirmationMode,
+		types.ConfirmationMode_SAFE,
 		types.WithEVMRevertOptions(event.RevertOptions),
 	)
 }
@@ -487,9 +484,6 @@ func (ob *Observer) newDepositAndCallInboundVote(event *gatewayevm.GatewayEVMDep
 		coinType = coin.CoinType_Gas
 	}
 
-	// determine confirmation mode
-	confirmationMode := ob.GetInboundConfirmationMode(event.Raw.BlockNumber)
-
 	return *types.NewMsgVoteInbound(
 		ob.ZetacoreClient().GetKeys().GetOperatorAddress().String(),
 		event.Sender.Hex(),
@@ -508,7 +502,7 @@ func (ob *Observer) newDepositAndCallInboundVote(event *gatewayevm.GatewayEVMDep
 		types.ProtocolContractVersion_V2,
 		false, // currently not relevant since calls are not arbitrary
 		types.InboundStatus_SUCCESS,
-		confirmationMode,
+		types.ConfirmationMode_SAFE,
 		types.WithEVMRevertOptions(event.RevertOptions),
 		types.WithCrossChainCall(true),
 	)